requirements.txt
noise==1.2.2
numpy==1.18.4
Pillow==7.1.2


